Split up synchronization hook
